requirements.txt
mysql-connector-python==8.0.30
protobuf==3.20.1
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
PyQt5Designer==5.14.1
